requirement.txt
aip==1.0.0.5
aircv==1.4.6
alembic==0.9.9
aniso8601==3.0.0
apkutils==0.3.0
appdirs==1.4.3
Appium-Python-Client==0.26
APScheduler==3.5.1
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
atx==1.2.1.dev5
bcrypt==3.1.4
beautifulsoup4==4.6.0
black==18.5b1
bleach==2.1.2
blinker==1.4
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cigam==0.0.3
click==6.7
colorama==0.3.9
configparser==3.5.0
coverage==4.5.1
cryptography==2.2.2
cssselect==1.0.3
cx-Freeze==5.1.1
cycler==0.10.0
ddt==1.1.2
decorator==4.3.0
docutils==0.14
enum34==1.1.6
facebook-wda==0.2.1
fake-useragent==0.1.10
Faker==0.8.16
fire==0.1.3
Flask==1.0.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
futures==3.0.5
html5lib==1.0.1
humanize==0.5.1
idna==2.6
imageio==2.3.0
itchat==1.3.10
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
kitchen==1.2.5
kiwisolver==1.0.1
Logbook==1.4.0
lxml==4.2.1
Mako==1.0.7
maproxy==0.0.12
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
more-itertools==4.2.0
namedlist==1.7
numpy==1.14.4
paramiko==2.4.1
parse==1.8.2
Pillow==5.1.0
pluggy==0.6.0
progress==1.3
py==1.5.3
pyasn1==0.4.3
pycparser==2.18
pyee==5.0.0
pyelftools==0.24
PyMySQL==0.8.1
PyNaCl==1.2.1
pyparsing==2.2.0
pypng==0.0.18
pyppeteer==0.0.17
PyQRCode==1.2.1
pyquery==1.4.0
pytest==3.6.0
pytest-allure-adaptor==1.7.10
python-dateutil==2.7.3
python-editor==1.0.3
python-hessian==1.0.2
pytz==2018.4
PyYAML==3.11
requests==2.18.4
requests-html==0.9.0
retry==0.9.2
robobrowser==0.5.3
robotframework==3.0.3
robotframework-appiumlibrary==1.4.6
robotframework-requests==0.4.7
robotframework-seleniumlibrary==3.1.1
selenium==3.11.0
six==1.11.0
span==0.0.1
SQLAlchemy==1.2.7
text-unidecode==1.2
tornado==5.0.2
tqdm==4.5.0
tzlocal==1.5.1
uiautomator2==0.1.2
urllib3==1.22
virtualenv==16.0.0
w3lib==1.19.0
webencodings==0.5.1
websockets==4.0.1
weditor==0.0.3
Werkzeug==0.14.1
whichcraft==0.4.1
XlsxWriter==1.0.5
xmltodict==0.11.0
yagmail==0.10.212
you-get==0.4.1060
